Added tests for expo-dev-menu AppDelegate plugin
