+++ packages/expo-dev-menu/package.json
@@ -41,6 +41,7 @@
     "preset": "expo-module-scripts"
   },
   "dependencies": {
+    "@expo/config-plugins": "^1.0.28",
     "expo-dev-menu-interface": "0.2.0"
   },
   "devDependencies": {
